% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{addLogs}
\alias{addLogs}
\title{Add Logs}
\usage{
addLogs(loga, logb)
}
\arguments{
\item{loga}{first number in log space.}

\item{logb}{first number in log space.}
}
\description{
Given log(a) and log(b), returns log(a + b)
}
