% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgbb.R
\name{bgbb.PosteriorMeanDropoutRate}
\alias{bgbb.PosteriorMeanDropoutRate}
\title{BG/BB Posterior Mean Dropout Rate}
\usage{
bgbb.PosteriorMeanDropoutRate(params, x, t.x, n.cal)
}
\arguments{
\item{params}{BG/BB parameters - a vector with alpha, beta, gamma, and
delta, in that order. Alpha and beta are unobserved parameters for the
beta-Bernoulli transaction process. Gamma and delta are unobserved
parameters for the beta-geometric dropout process.}

\item{x}{the number of repeat transactions made by the customer in the
calibration period. Can also be vector of frequencies - see details.}

\item{t.x}{recency - the transaction opportunity in which the customer made
their last transaction. Can also be a vector of recencies - see details.}

\item{n.cal}{number of transaction opportunities in the calibration
period. Can also be a vector of calibration period transaction
opportunities - see details.}
}
\value{
The posterior mean dropout rate.
}
\description{
Computes the mean value of the marginal posterior value of Theta, the geometric dropout process parameter.
}
\details{
E(Theta | alpha, beta, gamma, delta, x, t.x, n). This is calculated by setting \code{l = 0} and \code{m = 1} in \code{\link{bgbb.PosteriorMeanLmProductMoment}}.

\code{x}, \code{t.x}, and \code{n.cal} may be vectors. The standard rules for vector operations apply - if they are not of the same length, shorter vectors will be recycled (start over at the first element) until they are as long as the longest vector. It is advisable to keep vectors to the same length and to use single values for parameters that are to be the same for all calculations. If one of these parameters has a length greater than one, the output will be also be a vector.
}
\examples{
 
data(donationsSummary)

rf.matrix <- donationsSummary$rf.matrix
# donationsSummary$rf.matrix already has appropriate column names

# starting-point parameters
startingparams <- c(1, 1, 0.5, 3)
# estimated parameters
est.params <- bgbb.EstimateParameters(rf.matrix, startingparams)

# return the posterior mean dropout rate vector
bgbb.rf.matrix.PosteriorMeanDropoutRate(est.params, rf.matrix)
}
\references{
Fader, Peter S., Bruce G.S. Hardie, and Jen Shang.
"Customer-Base Analysis in a Discrete-Time Noncontractual Setting."
\emph{Marketing Science} 29(6), pp. 1086-1108. 2010. INFORMS.
\href{http://www.brucehardie.com/papers/020/}{Web.}
}
\seealso{
\code{\link{bgbb.rf.matrix.PosteriorMeanDropoutRate}}
}
