% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bgnbd.PlotDropoutRateHeterogeneity}
\alias{bgnbd.PlotDropoutRateHeterogeneity}
\title{BG/NBD Plot Dropout Probability Heterogeneity}
\usage{
bgnbd.PlotDropoutRateHeterogeneity(params, lim = NULL)
}
\arguments{
\item{params}{BG/NBD parameters - a vector with r, alpha, a, and b, in that order. r and alpha are unobserved parameters for the NBD transaction process. a and b are unobserved parameters for the Beta geometric dropout process.}

\item{lim}{upper-bound of the x-axis. A number is chosen by the
function if none is provided.}
}
\value{
Distribution of customers' probabilities of dropping out.
}
\description{
Plots and returns the estimated gamma distribution of p
(customers' probability of dropping out immediately after
a transaction).
}
\examples{
params <- c(0.243, 4.414, 0.793, 2.426)
bgnbd.PlotDropoutRateHeterogeneity(params)
params <- c(0.243, 4.414, 1.33, 2.426)
bgnbd.PlotDropoutRateHeterogeneity(params)
}

