% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bgnbd.PlotTrackingInc}
\alias{bgnbd.PlotTrackingInc}
\title{BG/NBD Tracking Incremental Transactions Comparison}
\usage{
bgnbd.PlotTrackingInc(params, T.cal, T.tot, actual.inc.tracking.data,
  xlab = "Week", ylab = "Transactions", xticklab = NULL,
  title = "Tracking Weekly Transactions")
}
\arguments{
\item{params}{BG/NBD parameters - a vector with r, alpha, a, and b, in that order. r and alpha are unobserved parameters for the NBD transaction process. a and b are unobserved parameters for the Beta geometric dropout process.}

\item{T.cal}{a vector to represent customers' calibration period
lengths (in other words, the "T.cal" column from a
customer-by-sufficient-statistic matrix).}

\item{T.tot}{end of holdout period. Must be a single value, not a
vector.}

\item{actual.inc.tracking.data}{vector containing the incremental
number of repeat transactions made by customers for each period in
the total time period (both calibration and holdout periods). See
details.}

\item{xlab}{descriptive label for the x axis.}

\item{ylab}{descriptive label for the y axis.}

\item{xticklab}{vector containing a label for each tick mark on the x axis.}

\item{title}{title placed on the top-center of the plot.}
}
\value{
Matrix containing actual and expected incremental repeat transactions.
}
\description{
Plots the actual and expected incremental total repeat
transactions by all customers for the calibration and holdout
periods, and returns this comparison in a matrix.
}
\details{
\code{actual.inc.tracking.data} does not have to be in the same
unit of time as the \code{T.cal} data. \code{T.tot} will
automatically be divided into periods to match the length of
\code{actual.inc.tracking.data}. See
\code{\link{bgnbd.ExpectedCumulativeTransactions}}.

The holdout period should immediately follow the calibration
period. This function assume that all customers' calibration
periods end on the same date, rather than starting on the same
date (thus customers' birth periods are determined using
\code{max(T.cal) - T.cal} rather than assuming that it is 0).
}
\examples{
data(cdnowSummary)
cal.cbs <- cdnowSummary$cbs
# cal.cbs already has column names required by method

# Cumulative repeat transactions made by all customers across calibration
# and holdout periods
cu.tracking <- cdnowSummary$cu.tracking
# make the tracking data incremental
inc.tracking <- dc.CumulativeToIncremental(cu.tracking)

# parameters estimated using bgnbd.EstimateParameters
est.params <- c(0.243, 4.414, 0.793, 2.426)

# All parameters are in weeks; the calibration period lasted 39
# weeks and the holdout period another 39.
bgnbd.PlotTrackingInc(est.params, T.cal=cal.cbs[,"T.cal"], T.tot=78, inc.tracking)
}

