% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdnowElog}
\alias{cdnowElog}
\title{CDNOW event log data}
\format{
A comma-delimited file representing an event log with 6,919 entries.
It has 5 columns: The customer's ID in the master dataset, the customer's
ID in this dataset (which represents 1/10th of the master dataset), the
date of the transaction in the format "\%Y\%m\%d" (e.g. 19970225), the number
of CDs purchased, and the dollar value of the transaction.
}
\source{
Can be found \href{https://www.brucehardie.com/datasets/}{online}.
}
\description{
Data representing the purchasing behavior of 2,357 CDNOW customers between
January 1997 and June 1998, in event log format.
}
\details{
The customers in this data represent 1/10th of the cohort of customers who
made their first transactions with CDNOW in the first quarter of 1997. CDNOW
was an online retailer, selling music and related products on the web since
1994.
}
\keyword{datasets}
