% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{dc.check.model.params}
\alias{dc.check.model.params}
\title{Check model params}
\usage{
dc.check.model.params(printnames, params, func)
}
\arguments{
\item{printnames}{Names to print parameter errors.}

\item{params}{Model parameters.}

\item{func}{Function calling dc.check.model.params.}
}
\value{
Stops the program if there is something wrong with the parameters.
}
\description{
Check model parameters for correctness.
}
