% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{discreteSimElog}
\alias{discreteSimElog}
\title{Discrete simulated annual event log data}
\format{
A comma-delimited file representing an event log with 52,432 entries.
It has 2 columns: The customer's ID and the date of the transaction in
standard R date format.
}
\description{
Data simulated using BG/BB model assumptions. Contains annual transaction
behavior for a period of 14 years, for a cohort of 10,000 customers who made
their first transactions in 1970.
}
\details{
This dataset was simulated in order to illustrate certain data-conversion
functions (see \code{\link{dc.MakeRFmatrixCal}}).
}
\keyword{datasets}
