% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnbd.R
\name{h2f1}
\alias{h2f1}
\title{Use Bruce Hardie's Gaussian hypergeometric implementation}
\usage{
h2f1(a, b, c, z)
}
\arguments{
\item{a, }{counterpart to A in \code{\link[hypergeo]{hypergeo}}}

\item{b, }{counterpart to B in \code{\link[hypergeo]{hypergeo}}}

\item{c, }{counterpart to C in \code{\link[hypergeo]{hypergeo}}}

\item{z, }{counterpart to z in \code{\link[hypergeo]{hypergeo}}}
}
\description{
In benchmarking \code{\link{pnbd.LL}} runs more quickly and
it returns the same results if it uses this helper instead of
\code{\link[hypergeo]{hypergeo}}, which is the default. But \code{h2f1}
is such a barebones function that in some edge cases it will keep
going until you get a segfault, where \code{\link[hypergeo]{hypergeo}}
would have failed with a proper error message.
}
\references{
Fader, Peter S., and Bruce G.S. Hardie. "A Note on Deriving the Pareto/NBD Model and
Related Expressions." November. 2005. Web. \url{http://www.brucehardie.com/notes/008/}
}
\seealso{
\code{\link[hypergeo]{hypergeo}}
}
