% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{subLogs}
\alias{subLogs}
\title{Subtract Logs}
\usage{
subLogs(loga, logb)
}
\arguments{
\item{loga}{first number in log space.}

\item{logb}{first number in log space.}
}
\description{
Given log(a) and log(b), returns log(a - b)
}
