% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addLogs}
\alias{addLogs}
\title{Add Logs}
\usage{
addLogs(loga, logb)
}
\arguments{
\item{loga}{first number in log space.}

\item{logb}{second number in log space.}
}
\description{
Given log(a) and log(b), returns log(a + b)
}
\keyword{internal}

