\name{cdnowElog}
\docType{data}
\alias{cdnowElog}
\title{CDNOW event log data}
\description{

  Data representing the purchasing behavior of 2357 CDNOW customers
  between January 1997 and June 1998, in event log format.
  
}
\format{
  
  A comma-delimited file representing an event log with 6919 entries. It
  has 5 columns: The customer's ID in the master dataset, the customer's
  ID in this dataset (which represents 1/10th of the master dataset), the
  date of the transaction in the format \dQuote{\%Y\%m\%d}
  (e.g. 19970225), the number of CDs purchased, and the dollar value of
  the transaction.

}

\source{

  Can be found online at \url{www.brucehardie.com/datasets}.
  
  }

\details{

  The customers in this data represent 1/10th of the cohort of customers who
  made their first transactions with CDNOW in the first quarter of
  1997. CDNOW was an online retailer, selling music and related products
  on the web since 1994.
  
}

\keyword{datasets}
