% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{dc.CumulativeToIncremental}
\alias{dc.CumulativeToIncremental}
\title{Cumulative to Incremental}
\usage{
dc.CumulativeToIncremental(cu)
}
\arguments{
\item{cu}{A vector containing cumulative transactions over time.}
}
\value{
Vector of incremental transactions.
}
\description{
Converts a vector of cumulative transactions to a vector of incremental transactions.
}
