% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dc.CumulativeToIncremental}
\alias{dc.CumulativeToIncremental}
\title{Cumulative to Incremental}
\usage{
dc.CumulativeToIncremental(cu)
}
\arguments{
\item{cu}{vector containing cumulative transactions over time.}
}
\value{
Vector of incremental transactions.
}
\description{
Converts a vector of cumulative transactions to a vector of
incremental transactions.
}

