% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{dc.WriteLine}
\alias{dc.WriteLine}
\title{Write Line}
\usage{
dc.WriteLine(...)
}
\arguments{
\item{...}{objects to print to the R console.}
}
\description{
Writes any number of arguments to the console.
}
\details{
The code is literally: cat(..., fill=TRUE); flush.console();
}
