% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spend.marginal.likelihood}
\alias{spend.marginal.likelihood}
\title{Gamma-gamma marginal likelihood}
\usage{
spend.marginal.likelihood(params, m.x, x)
}
\arguments{
\item{params}{a vector of gamma-gamma parameters: p, q, and gamma,
in that order. p is the shape parameter for each transaction. The
scale parameter for each transaction is distributed across
customers according to a gamma distribution with parameters q
(shape) and gamma (scale).}

\item{m.x}{the customer's average observed transaction value in
the calibration period. May also be a vector of average observed
transaction values - see details.}

\item{x}{the number of transactions the customer made in the
calibration period. May also be a vector of frequencies - see
details.}
}
\value{
The marginal distribution of a customer's average
transaction value. If \code{m.x} or \code{x} has a length greater
than 1, a vector of marginal likelihoods will be returned.
}
\description{
Calculates the marginal likelihood of a customer's average
transaction value.
}
\details{
\code{m.x} and \code{x} may be vectors. The standard rules for
vector operations apply - if they are not of the same length, the
shorter vector will be recycled (start over at the first element)
until it is as long as the longest vector. It is advisable to keep
vectors to the same length and to use single values for parameters
that are to be the same for all calculations. If one of these
parameters has a length greater than one, the output will be a
vector of probabilities.

This function will issue a warning if any of \code{m.x} or
\code{x} is 0, and will return a marginal likelihood of 0 for
those values.

f(m.x | p, q, gamma, x).
}
\examples{
params <- c(6, 4, 16)
# calculate the marginal distribution of the average transaction value
# of a customer who spent an average of $35 over 3 transactions.
spend.marginal.likelihood(params, m.x=35, x=3)

# Several values can also be computed at once:
spend.marginal.likelihood(params, m.x=30:40, x=3)
spend.marginal.likelihood(params, m.x=35, x=1:10)
spend.marginal.likelihood(params, m.x=30:40, x=1:11)
}
\references{
Fader, Peter S., Bruce G.S. Hardie, and Ka
L. Lee. \dQuote{RFM and CLV: Using Iso-Value Curves for Customer
Base Analysis.} \emph{Journal of Marketing Research} Vol.42,
pp.415-430. November. 2005. \url{http://www.brucehardie.com/papers.html}

See equation 3.
}

