% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subLogs}
\alias{subLogs}
\title{Subtract Logs}
\usage{
subLogs(loga, logb)
}
\arguments{
\item{loga}{first number in log space.}

\item{logb}{second number in log space.}
}
\description{
Given log(a) and log(b), returns log(a - b)
}
\keyword{internal}

